<header/Footer> Template 설정
