--- src/Main.scss
@@ -1,67 +1,0 @@
-// 페이지별 세부 스타일 설정
-
-#main-wrap {
-  font-family: 'Noto Sans KR', sans-serif;
-  color: #333;
-
-  header {
-    position: fixed;
-    top: 0;
-    .header-wrap {
-      display: flex;
-      flex-flow: row nowrap;
-      align-items: center;
-      width: 100vw;
-      height: 50px;
-      padding: 0 20px;
-      box-sizing: border-box;
-      background: rgba(255, 255, 255, 0.9);
-      box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-      h1 {
-        font-size: 18px;
-        margin-right: 10px;
-        padding-right: 10px;
-        border-right: 1px solid #ddd;
-      }
-      p {
-        font-size: 16px;
-      }
-    }
-  }
-  
-  body {
-    display: flex;
-    flex-flow: column nowrap;
-    align-items: center;
-    width: 100vw;
-    height: 100vh;
-    padding: 60px 15px 90px;
-    box-sizing: border-box;
-  }
-  
-  footer {
-    position: fixed;
-    bottom: 0;
-    .footer-wrap {
-      display: flex;
-      flex-flow: row nowrap;
-      align-items: center;
-      justify-content: flex-end;
-      width: 100vw;
-      height: 80px;
-      padding: 0 30px;
-      box-sizing: border-box;
-      background: #22252b;
-      color: #f5f5f5;
-      h1 {
-        font-size: 18px;
-        margin-right: 10px;
-        padding-right: 10px;
-        border-right: 1px solid #ddd;
-      }
-      p {
-        font-size: 15px;
-      }
-    }
-  }
-}
