+++ src/component/template/template.scss
@@ -1,0 +1,50 @@
+header {
+  position: fixed;
+  top: 0;
+  .header-wrap {
+    display: flex;
+    flex-flow: row nowrap;
+    align-items: center;
+    width: 100vw;
+    height: 50px;
+    padding: 0 20px;
+    box-sizing: border-box;
+    background: rgba(255, 255, 255, 0.9);
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    h1 {
+    font-size: 18px;
+    margin-right: 10px;
+    padding-right: 10px;
+    border-right: 1px solid #ddd;
+    }
+    p {
+    font-size: 16px;
+    }
+  }
+}
+
+footer {
+  position: fixed;
+  bottom: 0;
+  .footer-wrap {
+    display: flex;
+    flex-flow: row nowrap;
+    align-items: center;
+    justify-content: flex-end;
+    width: 100vw;
+    height: 80px;
+    padding: 0 30px;
+    box-sizing: border-box;
+    background: #22252b;
+    color: #f5f5f5;
+    h1 {
+      font-size: 18px;
+      margin-right: 10px;
+      padding-right: 10px;
+      border-right: 1px solid #ddd;
+    }
+    p {
+      font-size: 15px;
+    }
+  }
+}
