+++ src/component/template/template.tsx
@@ -1,28 +1,27 @@
 import React from 'react';
-import './Main.scss';
+import './template.scss';
 
-function Main() {
+interface IProps extends React.Props<{}> {
+}
+
+const Template = ({children}: IProps) => {
   return (
-    <div id='main-wrap'>
+    <>
       <header>
         <div className='header-wrap'>
           <h1>Lifelog</h1>
           <p>효동동 Blog</p>
         </div>
       </header>
-      <body>
-        <section className=''>
-          Node.js를 활용한 게시판 구현(예정)
-        </section>
-      </body>
+      {children}
       <footer>
         <div className='footer-wrap'>
           <h1>Lifelog</h1>
           <p>ⓒ jeong_hYoon. All rights reserved.</p>
         </div>
       </footer>
-    </div>
+    </>
   );
-}
+};
 
-export default Main;
+export default Template;
