+++ src/index.scss
@@ -1,4 +1,4 @@
-// 전체 페이지 스타일 기본설정
+// 전체-페이지 스타일 공통설정
 
 body {
   margin: 0;
