+++ src/pages/main/main.scss
@@ -1,0 +1,17 @@
+// 메인-페이지 세부 스타일 설정
+
+#main-wrap {
+  font-family: 'Noto Sans KR', sans-serif;
+  color: #333;
+  
+  body {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: center;
+    width: 100vw;
+    height: 100vh;
+    padding: 60px 15px 90px;
+    box-sizing: border-box;
+  }
+  
+}
