+++ src/pages/main/main.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import './main.scss';
+import Template from '../../component/template/template';
+
+function Main() {
+  return (
+    <div id='main-wrap'>
+      <Template>
+        <body>
+          <section className=''>
+            Node.js를 활용한 게시판 구현(예정)
+          </section>
+        </body>
+      </Template>
+    </div>
+  );
+}
+
+export default Main;
