+++ src/pages/sub/sub.scss
@@ -1,0 +1,6 @@
+// 서브-페이지 세부 스타일 설정
+
+#sub-wrap {
+  font-family: 'Noto Sans KR', sans-serif;
+  color: #333;
+}
