+++ src/pages/sub/sub.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import './sub.scss';
+
+function Sub() {
+  return (
+    <div id='sub-wrap'>
+    </div>
+  );
+}
+
+export default Sub;
